+++ lib/statistics.dart
@@ -16,5 +16,6 @@ export 'src/statistics_extension_num.dart';
 export 'src/statistics_extension_simd.dart';
 export 'src/statistics_forecast.dart';
 export 'src/statistics_metric.dart';
+export 'src/statistics_prime.dart';
 export 'src/statistics_tools.dart';
 export 'src/statistics_tools_csv.dart';